requirements.txt
streamlit
openai
python-dotenv
pypdf2
python-docx
pycryptodome